refactor(native): folderize resource wildlife and runtime contract tests
