+++ src/native/engine/runtime/player/player_registry.h
@@ -2,7 +2,7 @@
 #define BANANA_ENGINE_RUNTIME_PLAYER_REGISTRY_H
 
 #include "../../world/world.h"
-#include "../resource_domain.h"
+#include "../resource/resource_domain.h"
 #include <stdint.h>
 
 #ifdef __cplusplus
