+++ src/native/engine/runtime/resource/resource_domain.h
