+++ src/native/engine/runtime/wildlife/wildlife_gameplay.c
