+++ src/native/engine/runtime/wildlife/wildlife_gameplay.h
@@ -1,8 +1,8 @@
 #ifndef BANANA_ENGINE_RUNTIME_WILDLIFE_GAMEPLAY_H
 #define BANANA_ENGINE_RUNTIME_WILDLIFE_GAMEPLAY_H
 
-#include "../ai/controller.h"
-#include "../world/world.h"
+#include "../../ai/controller.h"
+#include "../../world/world.h"
 
 #ifdef __cplusplus
 extern "C"
