+++ tests/native/runtime/contracts/runtime_domain_adapters_test.c
@@ -1,5 +1,5 @@
 #include "runtime/merchant/merchant_trade_domain.h"
-#include "runtime/resource_domain.h"
+#include "runtime/resource/resource_domain.h"
 
 #include <stdio.h>
 #include <string.h>
